bot/requirements.txt
telebot==0.0.5
psycopg2-binary
requests==2.25.1
python-dotenv
translate
pandas
tabulate